[Velox] Move Fs.h to common/base
Move Fs.h out of common/base/tests to common/base
for non-test uses
