+++ velox/exec/tests/TableWriteTest.cpp
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "velox/common/base/Fs.h"
 #include "velox/dwio/common/DataSink.h"
 #include "velox/exec/tests/utils/HiveConnectorTestBase.h"
 #include "velox/exec/tests/utils/PlanBuilder.h"
-
-#include "velox/common/base/tests/Fs.h"
 
 using namespace facebook::velox;
 using namespace facebook::velox::exec;
